services/summary/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-generativeai==0.3.2
pydantic==2.5.0





